requirements.txt
streamlit==1.32.0
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.12.0
langchain-core==0.1.27
langchain-google-genai==0.0.9
google-generativeai==0.3.2
